@@ -1,4 +1,4 @@
-__version__ = '0.0.26'
+__version__ = '0.0.27'
 
 from ._accessory import *
 from ._blast import *
